double precision

+++ class/DeferredMaterial.h
@@ -9,7 +9,7 @@ public:
 	static std::shared_ptr<GBufferMaterial> CreateDefaultMaterial(bool skinned = false);
 
 public:
-	void Dispatch(const std::shared_ptr<CommandBuffer>& pCmdBuf, const Vector3f& groupNum, const Vector3f& groupSize, uint32_t pingpong = 0) override {}
+	void Dispatch(const std::shared_ptr<CommandBuffer>& pCmdBuf, const Vector3d& groupNum, const Vector3d& groupSize, uint32_t pingpong = 0) override {}
 	void Draw(const std::shared_ptr<CommandBuffer>& pCmdBuf, const std::shared_ptr<FrameBuffer>& pFrameBuffer, uint32_t pingpong = 0, bool overrideVP = false) override
 	{
 		DrawIndirect(pCmdBuf, pFrameBuffer, pingpong, overrideVP);
@@ -35,7 +35,7 @@ public:
 	static std::shared_ptr<DeferredShadingMaterial> CreateDefaultMaterial();
 
 public:
-	void Dispatch(const std::shared_ptr<CommandBuffer>& pCmdBuf, const Vector3f& groupNum, const Vector3f& groupSize, uint32_t pingpong = 0) override {}
+	void Dispatch(const std::shared_ptr<CommandBuffer>& pCmdBuf, const Vector3d& groupNum, const Vector3d& groupSize, uint32_t pingpong = 0) override {}
 	void Draw(const std::shared_ptr<CommandBuffer>& pCmdBuf, const std::shared_ptr<FrameBuffer>& pFrameBuffer, uint32_t pingpong = 0, bool overrideVP = false) override
 	{
 		DrawScreenQuad(pCmdBuf, pFrameBuffer, pingpong, overrideVP);
